% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.clean.R
\name{batch.clean}
\alias{batch.clean}
\title{Function that cleans up raw output of EJSCREEN batch processor}
\usage{
batch.clean(x, namesfile = "map_batch_to_friendly_fieldnames_2016.csv",
  oldcolnames, newcolnames)
}
\arguments{
\item{x}{Required. The output of the batch processor that runs EJSCREEN buffer report once per site.}

\item{namesfile}{Must specify either namesfile, or both oldcolnames and newcolnames. A csv filename, of file that maps fieldnames from those found in raw output of batch processor 
to more useful and friendly names that make more sense to me. Default was 'map_batch_to_friendly_fieldnames_2016.csv'
If function is called with the special value namesfile='keepnames' then the names are unchanged from those in x.}

\item{oldcolnames}{Optional. The names to be found in x, ignored if namesfile specified.}

\item{newcolnames}{Optional. The corresponding names to change them to, ignored if namesfile specified.}
}
\value{
the output is the output
}
\description{
Takes the raw output version of batch buffer results and cleans it up
  to make it ready for batch.summarize function
}
\author{
ejanalyst info@ejanalysis.com
}
