% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphelpers.R
\name{percent_map}
\alias{percent_map}
\title{Color-coded map of US counties (choropleth)}
\usage{
percent_map(
  var,
  color = "black",
  legend.title,
  min = 0,
  max = 100,
  projection = "polyconic"
)
}
\arguments{
\item{var}{vector of numbers to use for county map}

\item{color}{Default is 'black' - Color like 'blue' for \code{\link[grDevices]{colorRamp}} from white to color to generate a vector of gradations of fill colors for the map}

\item{legend.title}{text for legend}

\item{min}{Default is 0}

\item{max}{Default is 100}

\item{projection}{Default is 'polyconic' - for \link[maps]{map}}
}
\description{
Uses \code{\link[maps]{map}} function from maps package, to draw a map of US counties.
 
 PROBABLY WILL REPLACE THIS WITH \code{\link[choroplethr]{county_choropleth}} county_choropleth {choroplethr}
  or countypointmap in proxistat package
  
Note: percent map is designed to work with the counties data set
It will not work correctly with other data sets if their row order does
not exactly match the order in which the maps package plots counties.
It is hardcodedd to show quartiles as specific percentile bins, cutpoints in the legend.
}
