% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.parsed.R
\name{tab.parsed}
\alias{tab.parsed}
\title{tab.parsed is OBSOLETE}
\usage{
tab.parsed(myfile, folder = getwd(), skip = 2, wordsbeforefips.bg = 2,
  wordsbeforestate.bg = 3, wordsafterstate.bg = 5,
  wordsafterregion.bg = 2, wordsafterpop.bg = 0,
  wordsbeforemiles.ring = 0, wordsbeforestate.ring = 7,
  wordsafterstate.ring = 5, wordsafterregion.ring = 2,
  wordsafterpop.ring = 0, wordsbeforelon.ring = 6,
  wordsbeforelat.ring = 5)
}
\arguments{
\item{myfile}{name of file. a csv of saved tabular data from a buffer standard report in EJSCREEN 
and it will be imported as a data.frame plus info from the 1st two rows, using read.csv(file, as.is=TRUE, skip=2)}

\item{folder}{default is getwd()}

\item{skip}{default lines to skip is 2}

\item{wordsbeforefips.bg}{default=2}

\item{wordsbeforestate.bg}{default=3}

\item{wordsafterstate.bg}{default=5}

\item{wordsafterregion.bg}{default=2}

\item{wordsafterpop.bg}{default=0}

\item{wordsbeforemiles.ring}{=0}

\item{wordsbeforestate.ring}{=7}

\item{wordsafterstate.ring}{=5}

\item{wordsafterregion.ring}{=2}

\item{wordsafterpop.ring}{=0}

\item{wordsbeforelon.ring}{=6}

\item{wordsbeforelat.ring}{=5}
}
\value{
output is not described here
}
\description{
WAS TO READ IN THE CSV SAVED FROM THE SINGLE BUFFER REPORT SHOWN AS TABULAR VIEW ON SCREEN
  AND IT IS NOT USED BY THE BATCH SUMMARIZER. IT WAS USED TO READ ONE AT A TIME WHEN THAT WAS USEFUL A WHILE AGO.
}
