% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.compile.R
\name{tabs.compile}
\alias{tabs.compile}
\title{compiles into one table the several tables from files saved from EJSCREEN buffer reports - probably obsolete}
\usage{
tabs.compile(files, folder = getwd(), ...)
}
\arguments{
\item{files}{character vector of filenames with full paths}

\item{folder}{default is working directory, searches for all files in this folder unless files param given}

\item{...}{pass more to tabs.parsed}
}
\value{
the output is the output
}
\description{
This is a function to read a set of csv files saved from EJSCREEN tabular view and compile / format them
}
\author{
ejanalyst info@ejanalysis.com
}
