% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.reformat.R
\name{tabs.reformat}
\alias{tabs.reformat}
\title{reformats tabular data saved from a single buffer report - probably obsolete}
\usage{
tabs.reformat(fulltable, folder = getwd(), colcount = 17)
}
\arguments{
\item{fulltable}{input is fulltable, a set of tabular data from a buffer standard report in EJSCREEN}

\item{folder}{default is working directory}

\item{colcount}{colcount is 17 for circular buffer report, but differs for block group and maybe noncircular buffer reports}
}
\value{
the output is the output
}
\description{
based on code from make.finaltable.R
}
\author{
ejanalyst info@ejanalysis.com
}
