% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.pvalues.r
\name{wilcoxon.pvalues}
\alias{wilcoxon.pvalues}
\title{Wilcoxon Signed Rank Test p-values}
\usage{
wilcoxon.pvalues(x, n, digits = 4)
}
\arguments{
\item{x}{List of named quantiles (see example)}

\item{n}{Sample size (e.g., how many locations)}

\item{digits}{Default is to round to 4 digits}
}
\value{
Returns p values
}
\description{
See \url{https://en.wikipedia.org/wiki/Wilcoxon_signed-rank_test}
Uses \link{psignrank} function from stats package
}
\examples{
  x <- c(white=0, black=0, native=0, multi=0, minority=9, ling.isol=1, 
    age0.4=41, age0.17=0, age65.up=57, no.hs.diploma=41, low.income=48)
  # Sample Size (# of facilities w/ 3 mi. population >= 100):
  n <- 13
  print( wilcoxon.pvalues(x,n) ) 
}
